# perf: optimization of mapStakers function